+++ bidsme/Modules/MRI/_hmriNIFTI.py
@@ -41,7 +41,7 @@ metafields = {
             "InstitutionalDepartmentName": ("<InstitutionalDepartmentName>",
                                             None),
             "RepetitionTime": ("<round10:scale-3:RepetitionTime>", None),
-            "RepetitionTimeExitation": ("<round10:scale-3:RepetitionTime>",
+            "RepetitionTimeExcitation": ("<round10:scale-3:RepetitionTime>",
                                         None),
             "TaskName": ("<<bids:task>>", None)
             },